# perf(rust, python): fail fast on non-inferable strings in strptime if no `fmt` is provided