--- a/polars/polars-time/src/chunkedarray/utf8/infer.rs
+++ b/polars/polars-time/src/chunkedarray/utf8/infer.rs
@@ -18,7 +18,7 @@ const DATE_DMY_PATTERN: &str = r#"(?x)
         ['"]?            # optional quotes
         (?:\d{1,2})
         [-/]             # separator
-        (?:[01]?\d)
+        (?P<month>[01]?\d{1})
         [-/]             # separator
         (?:\d{4,})
         ['"]?            # optional quotes
@@ -32,7 +32,7 @@ const DATE_YMD_PATTERN: &str = r#"(?x)
         ['"]?            # optional quotes
         (?:\d{4,})
         [-/]             # separator
-        (?:[01]?\d)
+        (?P<month>[01]?\d{1})
         [-/]             # separator
         (?:\d{1,2})
         ['"]?            # optional quotes
@@ -46,7 +46,7 @@ const DATETIME_DMY_PATTERN: &str = r#"(?x)
         ['"]?  # optional quotes
         (?:\d{1,2})
         [-/]   # separator
-        (?:[01]?\d)
+        (?P<month>[01]?\d{1})
         [-/]   # separator
         (?:\d{4,})
         (?:
@@ -74,7 +74,7 @@ const DATETIME_YMD_PATTERN: &str = r#"(?x)
         ['"]?  # optional quotes
         (?:\d{4,})   # year
         [-/]   # separator
-        (?:[01]?\d)  # month
+        (?P<month>[01]?\d{1})  # month
         [-/]   # separator
         (?:\d{1,2})  # day
         (?:
@@ -101,7 +101,7 @@ const DATETIME_YMDZ_PATTERN: &str = r#"(?x)
         ['"]?  # optional quotes
         (?:\d{4,})
         [-/]  # separator
-        (?:[01]?\d)
+        (?P<month>[01]?\d{1})
         [-/]  # separator
         (?:\d{1,2})
         [T\ ]  # separator
@@ -134,11 +134,61 @@ static DATETIME_YMDZ_BYTES_RE: Lazy<BytesRegex> =
 impl Pattern {
     pub fn is_inferable(&self, val: &str) -> bool {
         match self {
-            Pattern::DateDMY => DATE_DMY_RE.is_match(val), //{
-            Pattern::DateYMD => DATE_YMD_RE.is_match(val), //{
-            Pattern::DatetimeDMY => DATETIME_DMY_RE.is_match(val),
-            Pattern::DatetimeYMD => DATETIME_YMD_RE.is_match(val),
-            Pattern::DatetimeYMDZ => DATETIME_YMDZ_RE.is_match(val),
+            Pattern::DateDMY => match DATE_DMY_RE.captures(val) {
+                Some(search) => (1..=12).contains(
+                    &search
+                        .name("month")
+                        .unwrap()
+                        .as_str()
+                        .parse::<u8>()
+                        .unwrap(),
+                ),
+                None => false,
+            },
+            Pattern::DateYMD => match DATE_YMD_RE.captures(val) {
+                Some(search) => (1..=12).contains(
+                    &search
+                        .name("month")
+                        .unwrap()
+                        .as_str()
+                        .parse::<u8>()
+                        .unwrap(),
+                ),
+                None => false,
+            },
+            Pattern::DatetimeDMY => match DATETIME_DMY_RE.captures(val) {
+                Some(search) => (1..=12).contains(
+                    &search
+                        .name("month")
+                        .unwrap()
+                        .as_str()
+                        .parse::<u8>()
+                        .unwrap(),
+                ),
+                None => false,
+            },
+            Pattern::DatetimeYMD => match DATETIME_YMD_RE.captures(val) {
+                Some(search) => (1..=12).contains(
+                    &search
+                        .name("month")
+                        .unwrap()
+                        .as_str()
+                        .parse::<u8>()
+                        .unwrap(),
+                ),
+                None => false,
+            },
+            Pattern::DatetimeYMDZ => match DATETIME_YMDZ_RE.captures(val) {
+                Some(search) => (1..=12).contains(
+                    &search
+                        .name("month")
+                        .unwrap()
+                        .as_str()
+                        .parse::<u8>()
+                        .unwrap(),
+                ),
+                None => false,
+            },
         }
     }
     pub fn is_inferable_bytes(&self, val: &[u8]) -> bool {