Benchmark:
```python
from IPython import get_ipython
ipython = get_ipython()

from datetime import datetime, date
import datetime as dt

import polars as pl
import io

dates = pl.date_range(datetime(1000, 1, 1), datetime(9999, 1, 1)).dt.strftime('%m-%d-%Y %H:%M:%S')
print(ipython.run_line_magic("timeit", "dates.str.strptime(pl.Datetime, strict=False)"))

data = 'my_column\\n' + '\\n'.join(dates)
print(ipython.run_line_magic("timeit", "pl.read_csv(io.StringIO(data), rechunk=False, try_parse_dates=True)"))

dates = pl.date_range(date(1000, 1, 1), date(9999, 1, 1)).dt.strftime('%m-%d-%Y')
print(ipython.run_line_magic("timeit", "dates.str.strptime(pl.Date, strict=False)"))

data = 'my_column\\n' + '\\n'.join(dates)
print(ipython.run_line_magic("timeit", "pl.read_csv(io.StringIO(data), rechunk=False, try_parse_dates=True)"))

dates = pl.Series(['2020-01-01'] + ['foo']*100_000)
print(ipython.run_line_magic("timeit", "dates.str.strptime(pl.Datetime, strict=False)"))

# current happy-paths
dates = pl.date_range(datetime(1000, 1, 1), datetime(9999, 1, 1)).dt.strftime('%d-%m-%Y %H:%M:%S')
print(ipython.run_line_magic("timeit", "dates.str.strptime(pl.Datetime, strict=False)"))

data = 'my_column\\n' + '\\n'.join(dates)
print(ipython.run_line_magic("timeit", "pl.read_csv(io.StringIO(data), rechunk=False, try_parse_dates=True)"))
```

Results: (first number is upstream/main, second one is this branch)
- 12.4 s ± 62 ms => 2.57 s ± 16.3 ms
- 5.36 s ± 212 ms => 2.15 s ± 10.5 ms
- 1.4 s ± 6.43 ms => 1.1 s ± 22.2 ms
- 759 ms ± 16.4 ms => 787 ms ± 18.8 ms
- 658 ms ± 13.4 ms => 25.1 ms ± 130 µs
- 1.25 s ± 3.54 ms => 1.25 s ± 13.2 ms
- 563 ms ± 4.73 ms => 552 ms ± 3.55 ms

https://www.kaggle.com/code/marcogorelli/polars-timing?scriptVersionId=125818935