@MarcoGorelli 2023-04-14
Right, after quite a few experiments, I have something which seems to work

I wasn't able to get any speedup by exiting early from `parse_bytes` so I left that one out for now