@ritchie46 2023-04-15
Is this one good to go?

Just to be sure, the pre-compiled bytes one always runs all patterns, and then the slow chrono parsers will check if it is inferrable right?