@MarcoGorelli 2023-04-15
yes, that's right - `parse_bytes` goes to polars' own parsers anyway, which have their own fail-fast conditions, like
```
        if val.len() - offset != (fmt_len as usize) {
            return None;
        }
```
(and I've added one for if the month is out-of-range), I think that's why it didn't help to try to exit early with a bytes regex

But for the chrono ones it seems to make a noticeable difference