@ritchie46 2023-04-15
Cool! Then this one is going in. Thanks a lot for this one. :raised_hands:  We could have horrific performance here. 

I want to add some benchmark tests in the future so we fail on regressions and hope we can tighten the timings more as well.